created components StartScreen ProductScreen CheckoutScreen
