+++ src/components/CheckoutScreen.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+/**
+ * Checkout page with chart, payment method and customer information
+ */
+
+ class CheckoutScreen extends React.Component {
+     render() {
+         return (
+             <></>
+         )
+     }
+ }
+
+ export default CheckoutScreen;
