+++ src/components/ProductScreen.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+/**
+ * Product page with dog and cat food
+ */
+
+ class ProductScreen extends React.Component {
+     render() {
+         return (
+             <></>
+         )
+     }
+ }
+
+ export default ProductScreen;
